TINY-11753: move `manageEnterInFigure` in EnterKey` file
