+++ modules/tinymce/src/core/main/ts/keyboard/EnterKey.ts
@@ -1,5 +1,6 @@
 import { Fun, Optional } from '@ephox/katamari';
 import { PlatformDetection } from '@ephox/sand';
+import { PredicateFind, SugarElement, SugarNode } from '@ephox/sugar';
 
 import Editor from '../api/Editor';
 import { EditorEvent } from '../api/util/EventDispatcher';
@@ -23,6 +24,21 @@ const handleEnterKeyEvent = (editor: Editor, event: EditorEvent<KeyboardEvent>) 
   editor.undoManager.transact(() => {
     InsertNewLine.insert(editor, event);
   });
+};
+
+const manageEnterInFigure = (editor: Editor) => {
+  const currentNode = SugarElement.fromDom(editor.selection.getNode());
+  if (SugarNode.isTag('figure')(currentNode)) {
+    PredicateFind.descendant(currentNode, SugarNode.isTag('figcaption'))
+      .filter((e) => !NodeType.isContentEditableFalse(e.dom))
+      .bind((figcaption) => PredicateFind.descendant(figcaption, SugarNode.isText))
+      .each((text) => {
+        const rng = editor.dom.createRng();
+        rng.setStart(text.dom, 0);
+        rng.setEnd(text.dom, 0);
+        editor.selection.setRng(rng);
+      });
+  }
 };
 
 const isCaretAfterKoreanCharacter = (rng: Range): boolean => {
@@ -71,6 +87,7 @@ const setup = (editor: Editor): void => {
       } else {
         handleEnterKeyEvent(editor, event);
       }
+      manageEnterInFigure(editor);
     }
   });
